Basic ut coverage to view model and interactor
The tests don't cover
the whole functionality but the basics. This entry also includes the
generation of mocks with Sourcery. One can regenerate the mocks by
calling `rake mocks` on ther root of the repository. Notice that I added
a `Rakefile` at the root of the project to do so.